Simplified validation of payload
